email/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
python-dotenv==1.0.0
fastapi-mail==1.2.6
